client/requirements_client.txt
requests
grpcio
grpcio-tools
click